resize/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# resize/resize_pad.py: 2
PyYAML == 5.1.1

# resize/resize_pad.py: 3
imutils == 0.5.2

# resize/resize_pad.py: 1
opencv_contrib_python == 4.1.0.25
